task 3 and 4 solved

+++ src/main/java/org/example/Book.java
@@ -8,7 +8,7 @@ import lombok.Data;
 public class Book {
     private String name;
     private String author;
-    private short publishingYear;
+    private int publishingYear;
     private String isbn;
     private  String publisher;
